requirements.txt
asgiref==3.8.1
boto3==1.34.145
botocore==1.34.145
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.8
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.0.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-storages==1.14.4
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
dropbox==12.0.2
gunicorn==22.0.0
idna==3.7
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
ply==3.11
psycopg==3.2.1
psycopg-binary==3.2.1
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
stone==3.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
